requirements.txt
# LA Geography - Python Dependencies
# Install with: uv pip install -r requirements.txt

# Core GIS libraries
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
ezesri>=0.2.0

# Data formats
pyarrow>=14.0.0  # For Parquet export

# Data fetching
requests>=2.31.0

# Census data
pygris>=0.1.6  # Census TIGER/Line shapefiles
census>=0.8.22  # Census API wrapper

# Configuration
pyyaml>=6.0.1

# Cloud storage
boto3>=1.34.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

